+++ contrib/git-changes-action/detector/doc.go
@@ -1,2 +1,0 @@
-// Package detector implements a detector for detecting changes in a git repo.
-package detector