+++ services/rfq/relayer/service/chainindexer.go
@@ -205,7 +205,7 @@ type decimalsRes struct {
 func (r *Relayer) handleDepositClaimed(ctx context.Context, event *fastbridge.FastBridgeBridgeDepositClaimed, chainID int) error {
 	err := r.inventory.Rebalance(ctx, chainID, event.Token)
 	if err != nil {
-		return fmt.Errorf("could not rebalance: %w", err)
+		logger.Errorf("could not rebalance: %v", err)
 	}
 	err = r.db.UpdateQuoteRequestStatus(ctx, event.TransactionId, reldb.ClaimCompleted)
 	if err != nil {